Refactor: pulled cbconfigparser to separate helper file
* parse_circuit_breaker_faildefault is new parser function
* Legacy parse_circuit_breaker still in crates/config/src/kdl/mod.rs
* TODO: Write tests and regress single_agent configuration
* TODO: migrate to new faildefault function if possible
